Switch to one term per year
The terms of senators are six years long in the Argentine Senate, but
only a third of those are relected every two years.  We've decided that
it will be easier to manage this if we artificially say that terms are
yearly, like the sessions of the National Congress of Argentina (of
which the senate is the upper house).