Scrape only :sort_name from members list
The scraper was not scraping the first names and given names correctly. The entire name was being stored in one, and nil was being assigned to the other.

We will scrape the name in first-name last-name order from the
individual member pages as only sort name is available from the member list page.