Scrape member name from MemberPage
We are sraping member data from the MembersList page but this does
not provide the member name in _first-name_ _last-name_ order.

In a previous commit, we created the MemberPage class. This commit
merges the data from that class with the MemberSection data so that
the hash for each member contains their name, given_name and
faily_name.